backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Mail==0.10.0
PyJWT==2.8.0
Werkzeug==3.0.3
pymongo==4.7.2
python-dotenv==1.0.1
gunicorn==22.0.0
